refactor(daemon): move ledger-migrate execution out of the action dispatcher
repo-cell-action-dispatch.ts is under the ontology-daemon-specialization
shrink-only gate (baseline 554); the SQLite generation branch pushed it to 581.
Both the generation-1 migration and the layout migration now live in
repo-cell-migration-actions.ts next to the other migration executors, and the
dispatcher keeps a one-line route (513 lines, -41 against baseline).
